Remove the splicing functions
